pyproject.toml
[project]
name = "omniparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic[bedrock,vertex]>=0.37.1",
    "azure-identity>=1.25.1",
    "boto3>=1.28.57",
    "dashscope>=1.25.5",
    "dill>=0.4.0",
    "easyocr>=1.7.2",
    "einops==0.8.0",
    "flash-attn @ file:///Users/lucas/works/OmniParserCustomized/flash_attn_dummy",
    "google-auth>=2,<3",
    "gradio>=6.2.0",
    "groq>=1.0.0",
    "jsonschema==4.22.0",
    "numpy==1.26.4",
    "openai==1.3.5",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "paddleocr<=2.10.0",
    "paddlepaddle>=3.2.2",
    "peft==0.17.1",
    "pre-commit==3.8.0",
    "pyautogui==0.9.54",
    "pytest==8.3.3",
    "pytest-asyncio==0.23.6",
    "rfdetr>=1.0.0",
    "ruff==0.6.7",
    "screeninfo>=0.8.1",
    "setuptools>=80.9.0",
    "streamlit>=1.38.0",
    "supervision==0.18.0",
    "timm>=1.0.22",
    "torchvision>=0.24.1",
    "transformers==4.49.0",
    "uiautomation>=2.0.29",
    "ultralytics==8.3.70",
]
